Implemented the first CPU instruction; `mov'
Of course required scaffolding code missing was discovered and added,
hence the amendment to so many classes.  If this is deemed as a clean
solution, we can work on implementing the next instruction.

The design is likely to change as the first instructions are added,
but will expectedly stabilize after it has been fiven enough
development time to mature.